refactor: use pure seed function
This makes sure that we can use it in a test setup and not as a CLI entrypoint
